+++ packages/orq-rc/RELEASES.md
@@ -3989,3 +3989,13 @@ Based on:
 - [typescript v3.13.0-rc.51] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.51] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.51 - packages/orq-rc
+
+## 2025-09-30 14:54:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.631.5 (2.716.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.52] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.52] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.52 - packages/orq-rc